requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
jinja2==3.1.2
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytz
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
PyJWT>=2.0.0
typing-extensions
boto3>=1.34.0
